Refactored TTTAttributed Cell Delegate from Comment View Controller. Part 4.
